run registry tests for both registries
